+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 На уже установленном 25.2.0: скачайте `@v25.2.1-prod` и выполните `sudo bash remnawave-manager.sh protocols`.
 
 SHA256:
-`67052bd1206712f1a2dd6a15bbfee66539b4813031628d93f02ea98c778d987c`
+`534353340dc8987375f5bc45242f01a97327c8d8713bfc5628f1884f891e7e60`
 
 ## 25.2.0-prod